Move test package and make INTERNAL_IP_RANGES private
